Issue #1394: Deprecate BrowserNavigationOverlay
